refactor: 공통 타입 분리 후 정리 및 HttpRepository 영향 받는 파일 수정
