+++ my-app/src/pages/CardAddPage/useCardSubmit.ts
@@ -1,8 +1,8 @@
 import { useNavigate } from 'react-router-dom';
 import { useAsync } from '../../components/common/commonHooks/useAsync';
-import { postCard } from '../../api/cardApi';
 import type { FormState } from './useCardFormState';
 import { sendingDataFormatter } from '../../utils/SendingDataFormatter';
+import { cardRepository } from '../../repository/RepositoryChanger';
 
 export const useCardSubmit = (formState: FormState, isFormValid: boolean) => {
   const navigate = useNavigate();
@@ -16,7 +16,7 @@ export const useCardSubmit = (formState: FormState, isFormValid: boolean) => {
     const sendingData = sendingDataFormatter(formState);
 
     try {
-      const response = await run(() => postCard(sendingData));
+      const response = await run(() => cardRepository.postCard(sendingData));
       if (!response) return;
 
       // 라우팅 페이지 연결
