+++ my-app/src/types/card.ts
@@ -1,0 +1,13 @@
+export interface SendingData {
+  number: string;
+  expirationDate: string;
+  cvc: string;
+  issuerCode: string;
+}
+
+export interface Card {
+  id: string;
+  issuerCode: string;
+  number: string;
+  expirationDate: string;
+}
