+++ my-app/src/utils/SendingDataFormatter.ts
@@ -1,9 +1,9 @@
-import type { SendingData } from '../api/cardApi';
 import {
   ISSUER_CODES,
   type CardCompany,
 } from '../components/cardCompanySection/CardCompanyConstants';
 import type { FormState } from '../pages/CardAddPage/useCardFormState';
+import type { SendingData } from '../types/card';
 
 export const sendingDataFormatter = (formState: FormState): SendingData => ({
   number: formState.cardNumber.join(''),
